\name{useroptfn-package}
\alias{useroptfn-package}
\alias{useroptfn}
\docType{package}
\title{
A set of functions to wrap user objective, gradient and hessian functions
}
\description{
Provides a wrapper around user functions for nonlinear optimization
to try to control for inadmissible arguments to user objective, gradient
or hessian functions, as well as provide for maximization.
}
\details{
\tabular{ll}{
Package: \tab useroptfn\cr
Version: \tab 2011-7.1\cr
Date: \tab 2011-07-01\cr
License: \tab GPL-2\cr
Lazyload: \tab Yes\cr
Repository: \tab R-Forge\cr
Repository/R-Forge/Project: \tab optimizer\cr
Repository/R-Forge/Revision: \tab 439\cr
Date/Publication: \tab 2011-07-07 00:48:49\cr
Packaged: \tab 2011-07-26 21:35:23 UTC; rforge\cr
Built: \tab R 2.13.1; ; 2011-07-27 13:12:25 UTC; unix\cr
}

Index:
\preformatted{
ufn                     Output tool for answers from optimx
ugr                     Output tool for answers from optimx
uhess                   Output tool for answers from optimx
}
}
\author{
John C Nash <nashjc@uottawa.ca>

Maintainer: John C Nash <nashjc@uottawa.ca>
}
\keyword{ package }
\seealso{
optimx
}

